Update hello.js
Some students may not bother using capital letters on the names of the celebrities, as long as they enter the correct names they should not be penalised for not using capital letters. The check will now alter the array contents so that both the expected and actual arrays are both lowercase.